requirements.txt
# LangChain and Document Processing
langchain>=0.1.0
langchain-community>=0.0.13
langchain_community>=0.0.13
langchain-huggingface>=0.0.13
langchain-openai>=0.0.5

# Vector Store
faiss-cpu>=1.8.0
chromadb>=0.4.22

# Document Loaders
unstructured>=0.11.8
markdown>=3.5.1

# Embeddings
openai>=1.7.2
tiktoken>=0.5.2

# Optional: Local embeddings (if you don't want to use OpenAI)
sentence-transformers>=2.2.2

# Chat Interface
gradio>=4.14.0
streamlit>=1.29.0

# Utilities
python-dotenv>=1.0.0

